% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fetch_gameDetail}
\alias{fetch_gameDetail}
\title{Fetch Game Detail}
\usage{
fetch_gameDetail(gameId)
}
\arguments{
\item{gameId}{A valid SIHF game ID. Use \code{get_results()} to rerieve
filtered lists of game summaries that include game IDs.}
}
\value{
httr response object
}
\description{
Fetch Game Detail
}
\examples{
gameDetail <- fetch_gameDetail(20171105078373)
}
