% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_alias_from_header}
\alias{get_alias_from_header}
\title{Get Field Alias from Header}
\usage{
get_alias_from_header(header)
}
\arguments{
\item{header}{Header element}
}
\value{
Vector of field aliases
}
\description{
Internal function to get alias vector from header element
}
