% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_gameDetail}
\alias{get_gameDetail}
\title{Get Game Detail}
\usage{
get_gameDetail(gameId)
}
\arguments{
\item{gameId}{A valid SIHF game ID. Use \code{get_results()} to rerieve
filtered lists of game summaries that include game IDs.}
}
\value{
httr response object
}
\description{
Get Game Detail
}
\examples{
gameDetail <- get_gameDetail(20171105078373)
}
