% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDetails.R
\name{parse.gameDetail.details}
\alias{parse.gameDetail.details}
\title{Parse Details}
\usage{
parse.gameDetail.details(gameDetail)
}
\arguments{
\item{gameDetail}{SIHF API gameDetail object (see \code{fetch_gameDetail()})}
}
\value{
Parsed content of \code{gameDetail$content$details}
}
\description{
Parse Details
}
