% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gameDetail_lineup.R
\name{parse.gameDetail.lineUp.field}
\alias{parse.gameDetail.lineUp.field}
\title{Parse Field Player Entry in Team Line-Up}
\usage{
parse.gameDetail.lineUp.field(lineUp.field)
}
\arguments{
\item{lineUp.field}{Field Player Entry in Team Line-Up}
}
\value{
Data frame with side, line and playerId
}
\description{
Parse Field Player Entry in Team Line-Up
}
