% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gameDetail_lineup.R
\name{parse.gameDetail.lineUp.goalkeepers}
\alias{parse.gameDetail.lineUp.goalkeepers}
\title{Parse Goalkeepers Entry in Team Line-Up}
\usage{
parse.gameDetail.lineUp.goalkeepers(lineUp.goalkeepers)
}
\arguments{
\item{lineUp.goalkeepers}{Goalkeepers Entry in Team Line-Up}
}
\value{
Data frame with line and playerId
}
\description{
Parse Goalkeepers Entry in Team Line-Up
}
