% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseLineup.R
\name{parse.gameDetail.lineUp}
\alias{parse.gameDetail.lineUp}
\title{Parse Line-Up}
\usage{
parse.gameDetail.lineUp(gameDetail)
}
\arguments{
\item{gameDetail}{SIHF API gameDetail object (see \code{fetch_gameDetail()})}
}
\value{
List with three elements: "gameId", "players" and "coach"
  with line-ups of both home and away team.
}
\description{
Parse Line-Up
}
\examples{
gameDetail <- fetch_gameDetail(20171105078373)
parse.gameDetail.lineUp(gameDetail)
}
