% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gameDetail_lineup.R
\name{parse.gameDetail.lineUp.team}
\alias{parse.gameDetail.lineUp.team}
\title{Parse Team Line-Up}
\usage{
parse.gameDetail.lineUp.team(lineUp.team)
}
\arguments{
\item{lineUp.team}{Team Line-Up}
}
\value{
A list with two elements "players" and "coach"
}
\description{
Parse Team Line-Up
}
