% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gameDetail_players.R
\name{parse.gameDetail.players}
\alias{parse.gameDetail.players}
\title{Parse Players}
\usage{
parse.gameDetail.players(gameDetail)
}
\arguments{
\item{gameDetail}{SIHF API gameDetail object (see \code{fetch_gameDetail()})}
}
\value{
Tibble containing player information. In particular, this matches teamId and jersey number to playerId.
}
\description{
Parse Players
}
\examples{
gameDetail <- fetch_gameDetail(20171105078373)
parse.gameDetail.players(gameDetail)
}
