% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gameDetail.R
\name{parse.gameDetail}
\alias{parse.gameDetail}
\title{Parse Game Detail}
\usage{
parse.gameDetail(gameDetail)
}
\arguments{
\item{gameDetail}{SIHF API gameDetail object (see \code{fetch_gameDetail()})}
}
\value{
Tibbles containing the parsed information from game detail. Moste information is nested in the list columns.
}
\description{
Parse game detail object as returned by \code{fetch_gameDetail()} into a set of flat tibbles.
}
\examples{
gameDetail <- fetch_gameDetail(20171105078373)
gameDetail.parsed <- parse.gameDetail(gameDetail)
}
