% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gameDetail_result.R
\name{parse.gameDetail.result}
\alias{parse.gameDetail.result}
\title{Parse Result}
\usage{
parse.gameDetail.result(gameDetail)
}
\arguments{
\item{gameDetail}{SIHF API gameDetail object (see \code{fetch_gameDetail()})}
}
\value{
Tibble containing parsed information of \code{gameDetail$content$result}
}
\description{
Parse Result
}
