% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gameDetail_stats.R
\name{parse.gameDetail.stats.players}
\alias{parse.gameDetail.stats.players}
\title{Parse Stats Players}
\usage{
parse.gameDetail.stats.players(players)
}
\arguments{
\item{players}{\code{gameDetail$content$stats[[1]]} and \code{gameDetail$content$stats[[3]]}}
}
\value{
Parsed goalie stats (one team / game)
}
\description{
Parse Stats Players
}
\examples{
gameDetail <- fetch_gameDetail(20171105078373)
parse.gameDetail.stats.players(gameDetail$content$stats[[1]])
}
