% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseStats.R
\name{parse.gameDetail.stats}
\alias{parse.gameDetail.stats}
\title{Parse Stats}
\usage{
parse.gameDetail.stats(gameDetail)
}
\arguments{
\item{gameDetail}{SIHF API gameDetail object (see \code{fetch_gameDetail()})}
}
\value{
List of three tibbles: players, goalies, teams
}
\description{
Parse Stats
}
\examples{
gameDetail <- fetch_gameDetail(20171105078373)
parse.gameDetail.stats(gameDetail)
}
