% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseStats.R
\name{parse.gameDetail.stats.teams}
\alias{parse.gameDetail.stats.teams}
\title{Parse Stats Team}
\usage{
parse.gameDetail.stats.teams(teams)
}
\arguments{
\item{teams}{\code{gameDetail$content$stats[[5]]}}
}
\value{
Parsed teams stats (one game, both teams)
}
\description{
Parse Stats Team
}
\examples{
gameDetail <- fetch_gameDetail(20171105078373)
parse.gameDetail.stats.teams(gameDetail$content$stats[[5]])
}
