% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSummary.R
\name{parse.gameDetail.summary}
\alias{parse.gameDetail.summary}
\title{Parse Summary}
\usage{
parse.gameDetail.summary(gameDetail)
}
\arguments{
\item{gameDetail}{SIHF API gameDetail object (see \code{fetch_gameDetail()})}
}
\value{
Parsed summary information from gameDetail
}
\description{
Parse Summary
}
\examples{
gameDetail <- fetch_gameDetail(20171105078373)
parse.gameDetail.summary(gameDetail)
}
