% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHeader.R
\name{parse.header}
\alias{parse.header}
\title{Parse Header}
\usage{
parse.header(gameDetail)
}
\arguments{
\item{gameDetail}{SIHF API gameDetail object (see \code{fetch_gameDetail()})}
}
\value{
Tibble
}
\description{
Parse Header
}
\examples{
gameDetail <- fetch_gameDetail(20171105078373)
parse.header(gameDetail)
}
