% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.R
\name{parse.results.item}
\alias{parse.results.item}
\title{Results Single Item}
\usage{
parse.results.item(x)
}
\arguments{
\item{x}{Single item of \code{results$content$data}}
}
\value{
Tibble with parsed content of x
}
\description{
Results Single Item
}
