% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.R
\name{parse.results.item.scorePerPeriod.team}
\alias{parse.results.item.scorePerPeriod.team}
\title{Pares Results Single Item ScorePerPeriod Team}
\usage{
parse.results.item.scorePerPeriod.team(x)
}
\arguments{
\item{x}{\code{results$content$data$scorePerPeriod$homeTeam} or \code{results$content$data$scorePerPeriod$homeTeam}}
}
\value{
Tibble with parsed content of x
}
\description{
Pares Results Single Item ScorePerPeriod Team
}
