% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_results.R
\name{parse.results}
\alias{parse.results}
\title{Parse Results}
\usage{
parse.results(results)
}
\arguments{
\item{results}{"Results" output of SIHF Stats API as retrieved by \code{fetch_results()}}
}
\value{
A tibble with the parsed content of results output
}
\description{
Parse Results
}
\examples{
results_raw <- fetch_results(c(Season = 2017))
print(results_raw)

results <- parse.results(results_raw)
}
