% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseStats.R
\name{parse.stats.goalies}
\alias{parse.stats.goalies}
\title{Parse Stats Goalies}
\usage{
parse.stats.goalies(goalies)
}
\arguments{
\item{goalies}{\code{gameDetail$content$stats[[2]]} and \code{gameDetail$content$stats[[4]]}}
}
\value{
Parsed goalie stats (one team / game)
}
\description{
Parse Stats Goalies
}
\examples{
gameDetail <- fetch_gameDetail(20171105078373)
parse.stats.goalies(gameDetail$content$stats[[2]])
}
