% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSummary.R
\name{parse.summary.period.child}
\alias{parse.summary.period.child}
\title{Parse Summary Period Element}
\usage{
parse.summary.period.child(x)
}
\arguments{
\item{x}{Summary period element ('fouls' or 'goals')}
}
\value{
Parsed tibble of information contained by summary period element
}
\description{
Parse Summary Period Element
}
