% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSummary.R
\name{parse.summary.period}
\alias{parse.summary.period}
\title{Parse Summary Period}
\usage{
parse.summary.period(period)
}
\arguments{
\item{period}{Single summary period element}
}
\value{
List of tibbles containing information in summary period elements
}
\description{
Parse Summary Period
}
