% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSummary.R
\name{parse.summary.periods}
\alias{parse.summary.periods}
\title{Parse Summary Periods}
\usage{
parse.summary.periods(periods)
}
\arguments{
\item{periods}{List of summary period elements}
}
\value{
List of tibbles 'fouls' and 'goals' across all periods
}
\description{
Parse Summary Periods
}
