% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_percent}
\alias{parse_percent}
\title{Parse Percent}
\usage{
parse_percent(x, ...)
}
\arguments{
\item{x}{A character vector of percent numbers}

\item{...}{Further parameters passed to \code{parse_number}}
}
\value{
Numeric vector
}
\description{
Returns a numeric fraction instead from character percentage.
}
