% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.sihf_api}
\alias{print.sihf_api}
\title{Print Object of Class 'sihf_api'}
\usage{
\method{print}{sihf_api}(x, ...)
}
\arguments{
\item{x}{Return object of \code{sifh_api()}}

\item{...}{further arguments passed to or from other methods.}
}
\value{
x (invisibly)
}
\description{
Print Object of Class 'sihf_api'
}
