% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ut_firstToUpper}
\alias{ut_firstToUpper}
\title{First To Upper}
\usage{
ut_firstToUpper(x)
}
\arguments{
\item{x}{A string vector}
}
\value{
String vector with the first character of every element converte to upper case.
}
\description{
Convert first character of string to upper case (leaving all others as is).
}
\examples{
x <- "some string"
ut_firstToUpper(x)
}
